requirements.txt
# Core ML/DL libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.41.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.7.0

# Feature extraction and NLP
nltk>=3.8.0
textblob>=0.17.0
wordcloud>=1.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Progress bars and utilities
tqdm>=4.65.0
rich>=13.0.0

# Configuration and logging
PyYAML>=6.0
omegaconf>=2.3.0
wandb>=0.15.0

# Web scraping and requests
requests>=2.31.0
urllib3>=1.26.0

# Parallel processing
joblib>=1.3.0

# File handling
h5py>=3.9.0
pyarrow>=12.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Jupyter notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.0

# Optional: GPU monitoring
gpustat>=1.1.0
py3nvml>=0.2.7

# Optional: Model optimization
optimum>=1.12.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Optional: Advanced metrics
torchmetrics>=1.0.0

# Optional: Hyperparameter optimization
optuna>=3.3.0
ray[tune]>=2.6.0

# Memory and system monitoring
psutil>=5.9.0

# Enhanced ML algorithms
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Enhanced image processing
timm>=0.9.0

# Optional: Distributed training  
deepspeed>=0.10.0

# Optional: Model serving
fastapi>=0.100.0
uvicorn>=0.23.0

# Version constraints for compatibility
# Ensure compatible versions
tokenizers>=0.13.0,<0.15.0
safetensors>=0.3.0
huggingface-hub>=0.16.0